+++ noodles-sam/src/io/reader/record_buf/cigar/op/kind.rs
@@ -1,3 +1,5 @@
+//! SAM record buf CIGAR operation kind reader.
+
 use std::{error, fmt};
 
 use crate::alignment::record::cigar::op::Kind;
@@ -8,7 +10,7 @@ pub enum ParseError {
     /// Unexpected EOF.
     UnexpectedEof,
     /// The input is invalid.
-    Invalid { actual: u8 },
+    Invalid(u8),
 }
 
 impl error::Error for ParseError {}
@@ -17,7 +19,7 @@ impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::UnexpectedEof => write!(f, "unexpected EOF"),
-            Self::Invalid { actual } => write!(
+            Self::Invalid(actual) => write!(
                 f,
                 "invalid input: expected {{M, I, D, N, S, H, P, =, X}}, got {c}",
                 c = char::from(*actual)
@@ -41,7 +43,7 @@ pub(super) fn parse_kind(src: &mut &[u8]) -> Result<Kind, ParseError> {
         b'P' => Ok(Kind::Pad),
         b'=' => Ok(Kind::SequenceMatch),
         b'X' => Ok(Kind::SequenceMismatch),
-        _ => Err(ParseError::Invalid { actual: *n }),
+        _ => Err(ParseError::Invalid(*n)),
     }
 }
 
@@ -66,10 +68,6 @@ mod tests {
         t(b"X", Kind::SequenceMismatch);
 
         assert_eq!(parse_kind(&mut &[][..]), Err(ParseError::UnexpectedEof));
-
-        assert_eq!(
-            parse_kind(&mut &b"!"[..]),
-            Err(ParseError::Invalid { actual: b'!' })
-        );
+        assert_eq!(parse_kind(&mut &b"!"[..]), Err(ParseError::Invalid(b'!')));
     }
 }