+++ noodles-sam/src/io/reader/record_buf/data.rs
@@ -1,4 +1,6 @@
-pub(crate) mod field;
+//! SAM record buf data reader.
+
+pub mod field;
 
 use std::{error, fmt};