+++ noodles-sam/src/io/reader/record_buf/data/field/tag.rs
@@ -1,3 +1,5 @@
+//! SAM record buf data field tag reader.
+
 use std::{error, fmt};
 
 use crate::alignment::record::data::field::Tag;