+++ noodles-sam/src/io/reader/record_buf/data/field/value/array.rs
@@ -1,4 +1,6 @@
-mod subtype;
+//! SAM record buf data field array value reader.
+
+pub mod subtype;
 
 use std::{error, fmt};