+++ noodles-sam/src/io/reader/record_buf/template_length.rs
@@ -1,3 +1,5 @@
+//! SAM record buf template length reader.
+
 use std::{error, fmt};
 
 /// An error when raw SAM record flags fail to parse.